caldav_util.c: move caldav_get_validators(), caldav_store_resource(), caldav_record_to_ical(), and assorted support functions from http_caldav.c to caldav_support.c
